+++ src/cmd/cmd_conc.mli
@@ -19,7 +19,7 @@ val cmd :
   -> solver:Smtml.Solver_type.t
   -> files:Fpath.t list
   -> profile:Fpath.t option
-  -> model_output_format:Cmd_utils.model_output_format
+  -> model_format:Cmd_utils.model_format
   -> entry_point:string option
   -> invoke_with_symbols:bool
   -> unit Result.t