+++ src/cmd/cmd_cpp.ml
@@ -97,7 +97,7 @@ let compile ~workspace ~entry_point ~includes ~opt_lvl ~out_file debug
 let cmd ~debug ~arch:_ ~workers ~opt_lvl ~includes ~files ~profiling ~unsafe
   ~optimize ~no_stop_at_failure ~no_value ~no_assert_failure_expression_printing
   ~deterministic_result_order ~fail_mode ~concolic ~solver ~profile
-  ~model_output_format ~entry_point ~invoke_with_symbols ~out_file ~workspace :
+  ~model_format ~entry_point ~invoke_with_symbols ~out_file ~workspace :
   unit Result.t =
   let* workspace =
     match workspace with
@@ -118,4 +118,4 @@ let cmd ~debug ~arch:_ ~workers ~opt_lvl ~includes ~files ~profiling ~unsafe
     ~profiling ~debug ~unsafe ~rac:false ~srac:false ~optimize ~workers
     ~no_stop_at_failure ~no_value ~no_assert_failure_expression_printing
     ~deterministic_result_order ~fail_mode ~workspace ~solver ~files ~profile
-    ~model_output_format ~entry_point ~invoke_with_symbols
+    ~model_format ~entry_point ~invoke_with_symbols