+++ src/cmd/cmd_replay.ml
@@ -154,11 +154,16 @@ let run_file ~unsafe ~optimize ~entry_point ~invoke_with_symbols filename model
   c
 
 let cmd ~profiling ~debug ~unsafe ~optimize ~replay_file ~source_file
-  ~entry_point ~invoke_with_symbols =
+  ~entry_point ~invoke_with_symbols ~model_format =
   if profiling then Log.profiling_on := true;
   if debug then Log.debug_on := true;
+  let parse_fn =
+    match model_format with
+    | Cmd_utils.Scfg -> Smtml.Model.Parse.Scfg.from_file
+    | Json -> Smtml.Model.Parse.Json.from_file
+  in
   let* model =
-    match Smtml.Model.Parse.Scfg.from_file replay_file with
+    match parse_fn replay_file with
     | Error (`Msg msg) -> Error (`Invalid_model msg)
     | Ok model ->
       let bindings = Smtml.Model.get_bindings model in